// Frontend/src/api/models/UsuarioOutput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jesus Antona
 * Ingeniería Web
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface UsuarioOutput
 */
export interface UsuarioOutput {
    /**
     * 
     * @type {string}
     * @memberof UsuarioOutput
     */
    idUser: string;
    /**
     * 
     * @type {string}
     * @memberof UsuarioOutput
     */
    usuario: string;
    /**
     * 
     * @type {string}
     * @memberof UsuarioOutput
     */
    nombre: string;
    /**
     * 
     * @type {string}
     * @memberof UsuarioOutput
     */
    apellido1: string;
    /**
     * 
     * @type {string}
     * @memberof UsuarioOutput
     */
    apellido2: string | null;
    /**
     * 
     * @type {number}
     * @memberof UsuarioOutput
     */
    valoracion: number;
    /**
     * 
     * @type {boolean}
     * @memberof UsuarioOutput
     */
    esAdmin: boolean;
}

/**
 * Check if a given object implements the UsuarioOutput interface.
 */
export function instanceOfUsuarioOutput(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "idUser" in value;
    isInstance = isInstance && "usuario" in value;
    isInstance = isInstance && "nombre" in value;
    isInstance = isInstance && "apellido1" in value;
    isInstance = isInstance && "apellido2" in value;
    isInstance = isInstance && "valoracion" in value;
    isInstance = isInstance && "esAdmin" in value;

    return isInstance;
}

export function UsuarioOutputFromJSON(json: any): UsuarioOutput {
    return UsuarioOutputFromJSONTyped(json, false);
}

export function UsuarioOutputFromJSONTyped(json: any, ignoreDiscriminator: boolean): UsuarioOutput {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'idUser': json['idUser'],
        'usuario': json['usuario'],
        'nombre': json['nombre'],
        'apellido1': json['apellido1'],
        'apellido2': json['apellido2'],
        'valoracion': json['valoracion'],
        'esAdmin': json['esAdmin'],
    };
}

export function UsuarioOutputToJSON(value?: UsuarioOutput | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'idUser': value.idUser,
        'usuario': value.usuario,
        'nombre': value.nombre,
        'apellido1': value.apellido1,
        'apellido2': value.apellido2,
        'valoracion': value.valoracion,
        'esAdmin': value.esAdmin,
    };
}