// Frontend/src/api/models/ProductoOutput.ts
export interface ProductoOutput {
    /**
     * 
     * @type {string}
     * @memberof ProductoOutput
     */
    idProd: string;
    /**
     * 
     * @type {string}
     * @memberof ProductoOutput
     */
    vendedor: string;
    /**
     * 
     * @type {string}
     * @memberof ProductoOutput
     */
    titulo: string;
    /**
     * 
     * @type {string}
     * @memberof ProductoOutput
     */
    descripcion: string;
    /**
     * 
     * @type {number}
     * @memberof ProductoOutput
     */
    lat: number;
    /**
     * 
     * @type {number}
     * @memberof ProductoOutput
     */
    _long: number;
    /**
     * 
     * @type {string}
     * @memberof ProductoOutput
     */
    direccion: string;
    /**
     * 
     * @type {number}
     * @memberof ProductoOutput
     */
    edad: number | null;
    /**
     * 
     * @type {number}
     * @memberof ProductoOutput
     */
    precio: number;
    /**
     * 
     * @type {number}
     * @memberof ProductoOutput
     */
    peso: number;
    /**
     * 
     * @type {string}
     * @memberof ProductoOutput
     */
    fechaPublicacion: string;
    /**
     * 
     * @type {string}
     * @memberof ProductoOutput
     */
    fechaCierre: string;
    /**
     * 
     * @type {boolean}
     * @memberof ProductoOutput
     */
    pagado: boolean;
    /**
     * 
     * @type {boolean}
     * @memberof ProductoOutput
     */
    valorado: boolean;
    /**
     * 
     * @type {Array<string>}
     * @memberof ProductoOutput
     */
    imagenes: Array<string> | null;
}